+++ ui-admin/src/portal/siteContent/HtmlSectionEditor.tsx
@@ -1,27 +1,25 @@
-import React, { useEffect, useId, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import {
-  FaqQuestion,
   HtmlSection,
-  SectionConfig,
-  SectionType,
-  socialMediaSites,
-  StepConfig,
-  SubGrid
+  SectionType
 } from '@juniper/ui-core'
 import Select from 'react-select'
-import { Button, IconButton } from 'components/forms/Button'
-import { faChevronDown, faChevronUp, faPlus, faTimes } from '@fortawesome/free-solid-svg-icons'
+import { IconButton } from 'components/forms/Button'
+import { faChevronDown, faChevronUp, faTimes } from '@fortawesome/free-solid-svg-icons'
 import { sectionTemplates } from './sectionTemplates'
 import classNames from 'classnames'
-import { TextInput } from '../../components/forms/TextInput'
-import { Textarea } from '../../components/forms/Textarea'
-import { Checkbox } from '../../components/forms/Checkbox'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { getMediaUrl, SiteMediaMetadata } from '../../api/api'
-import useReactSingleSelect from '../../util/react-select-utils'
-import { ImageConfig, MediaConfig } from '@juniper/ui-core/build/participant/landing/ConfiguredMedia'
-import { ButtonConfig } from '@juniper/ui-core/build/participant/landing/ConfiguredButton'
+import { SiteMediaMetadata } from 'api/api'
 import { PortalEnvContext } from '../PortalRouter'
+import { ButtonEditor } from './designer/components/ButtonEditor'
+import { FrequentlyAskedQuestionEditor } from './designer/components/FrequentlyAskedQuestionEditor'
+import { TitleEditor } from './designer/components/TitleEditor'
+import { BlurbEditor } from './designer/components/BlurbEditor'
+import { ImageEditor } from './designer/components/ImageEditor'
+import { StyleEditor } from './designer/components/StyleEditor'
+import { ParticipationStepsEditor } from './designer/components/ParticipantStepsEditor'
+import { LogoEditor } from './designer/components/LogoEditor'
+import { PhotoBioEditor } from './designer/components/PhotoBioEditor'
+import { SocialMediaEditor } from './designer/components/SocialMediaEditor'
 
 const SECTION_TYPES = [
   { label: 'FAQ', value: 'FAQ' },
@@ -34,13 +32,6 @@ const SECTION_TYPES = [
   { label: 'RAW_HTML', value: 'RAW_HTML' },
   { label: 'LINK_SECTIONS_FOOTER', value: 'LINK_SECTIONS_FOOTER' },
   { label: 'BANNER_IMAGE', value: 'BANNER_IMAGE' }
-]
-
-const BUTTON_TYPES = [
-  { label: 'Internal Link', value: 'internalLink' },
-  { label: 'External Link', value: 'externalLink' },
-  { label: 'Join Study', value: 'join' },
-  { label: 'Join Mailing List', value: 'mailingList' }
 ]
 
 /**
@@ -177,127 +168,6 @@ const HtmlSectionEditor = ({
   </>
 }
 
-const StyleOptions = ({ section, updateSection }: {
-  section: HtmlSection, updateSection: (section: HtmlSection) => void
-}) => {
-  const config = JSON.parse(section.sectionConfig || '{}') as SectionConfig
-  const contentId = useId()
-  const targetSelector = `#${contentId}`
-  const imagePositionOptions = [{ label: 'Left', value: 'left' }, { label: 'Right', value: 'right' }]
-  return (
-    <div>
-      <div className="pb-1">
-        <button
-          aria-controls={targetSelector}
-          aria-expanded="true"
-          className={classNames('btn w-100 py-2 px-0 d-flex text-decoration-none')}
-          data-bs-target={targetSelector}
-          data-bs-toggle="collapse"
-        >
-          <span className={'form-label fw-semibold mb-0'}>Style Options</span>
-          <span className="text-center px-2">
-            <FontAwesomeIcon icon={faChevronDown} className="hidden-when-collapsed"/>
-            <FontAwesomeIcon icon={faChevronUp} className="hidden-when-expanded"/>
-          </span>
-        </button>
-      </div>
-      <div className="collapse hide rounded-3 mb-2" id={contentId}
-        style={{ backgroundColor: '#eee', padding: '0.75rem' }}>
-        <TextInput label="Background Color" value={config.background as string}
-          placeholder={'Enter a value to override default'}
-          onChange={value => {
-            updateSection({ ...section, sectionConfig: JSON.stringify({ ...config, background: value }) })
-          }}/>
-        <TextInput label="Text Color" value={config.color as string}
-          placeholder={'Enter a value to override default'}
-          onChange={value => {
-            updateSection({ ...section, sectionConfig: JSON.stringify({ ...config, color: value }) })
-          }}/>
-        { Object.hasOwnProperty.call(config, 'image') && <div className='mt-2'><Checkbox label={'Full Width'}
-          checked={config.fullWidth as boolean == undefined ? false : config.fullWidth as boolean}
-          onChange={value => {
-            updateSection({ ...section, sectionConfig: JSON.stringify({ ...config, fullWidth: value }) })
-          }}/></div>}
-        { Object.hasOwnProperty.call(config, 'image') &&
-          <div className='mt-2'>
-            <label className='form-label fw-semibold'>Image Position</label>
-            <Select options={imagePositionOptions}
-              value={config.imagePosition ? imagePositionOptions.find(opt => opt.value === config.imagePosition)
-                : undefined}
-              onChange={opt => {
-                updateSection({ ...section, sectionConfig: JSON.stringify({ ...config, imagePosition: opt?.value }) })
-              }}/></div>}
-      </div>
-    </div>
-  )
-}
-
-const SocialMediaSectionEditor = ({ section, updateSection }: {
-  section: HtmlSection, updateSection: (section: HtmlSection) => void
-}) => {
-  const config = JSON.parse(section.sectionConfig || '{}') as SectionConfig
-  return (
-    <div>
-      {socialMediaSites.map(site => {
-        const handleKey = `${site.label.toLowerCase()}Handle`
-        return (
-          <div key={handleKey} className="d-flex align-items-center mb-2">
-            <div className="col-3">
-              <label className='fw-bold'>{site.label}</label>
-            </div>
-            <div className="col d-flex align-items-right col">
-              <TextInput value={config[handleKey] as string} placeholder={`Enter ${site.label} handle. ` +
-                  'Leave blank to hide'} onChange={value => {
-                updateSection({ ...section, sectionConfig: JSON.stringify({ ...config, [handleKey]: value }) })
-              }}/>
-            </div>
-          </div>
-        )
-      })}
-    </div>
-  )
-}
-
-const ImageSelector = ({ portalEnvContext, imageList, image, onChange }: {
-  portalEnvContext: PortalEnvContext,
-  imageList: SiteMediaMetadata[], image: ImageConfig, onChange: (image: SiteMediaMetadata) => void
-}) => {
-  const initial = imageList.find(media => media.cleanFileName === image.cleanFileName)
-  const [, setSelectedImage] = useState<SiteMediaMetadata | undefined>(initial)
-
-  const imageOptionLabel = (image: SiteMediaMetadata, portalShortcode: string) => <div>
-    {image.cleanFileName} <img style={{ maxHeight: '1.5em' }}
-      src={getMediaUrl(portalShortcode, image!.cleanFileName, image!.version)}/>
-  </div>
-
-  const {
-    onChange: imageOnChange, options, selectedOption, selectInputId
-  } = useReactSingleSelect(
-    imageList,
-    (media: SiteMediaMetadata) => ({ label: imageOptionLabel(media, portalEnvContext.portal.shortcode), value: media }),
-    setSelectedImage,
-    initial
-  )
-
-  return (
-    <div>
-      <Select
-        placeholder={'Select an image'}
-        isSearchable={false}
-        inputId={selectInputId}
-        options={options}
-        value={selectedOption}
-        onChange={opt => {
-          if (opt != undefined) {
-            imageOnChange(opt)
-            onChange(opt.value)
-          }
-        }}
-      />
-    </div>
-  )
-}
-
 const DynamicSectionEditor = ({ portalEnvContext, section, updateSection, siteMediaList }: {
   portalEnvContext: PortalEnvContext,
   section: HtmlSection, updateSection: (section: HtmlSection) => void, siteMediaList: SiteMediaMetadata[]
@@ -315,19 +185,28 @@ const DynamicSectionEditor = ({ portalEnvContext, section, updateSection, siteMe
 
   return (
     <div>
-      {hasTitle && <TitleEditor section={section} updateSection={updateSection}/>}
-      {hasBlurb && <BlurbEditor section={section} updateSection={updateSection}/>}
-      {hasImage && <ImageEditor portalEnvContext={portalEnvContext} section={section}
-        updateSection={updateSection} siteMediaList={siteMediaList}/>}
-      <StyleOptions section={section} updateSection={updateSection}/>
-      {hasSteps && <StepEditor portalEnvContext={portalEnvContext} section={section}
-        updateSection={updateSection} siteMediaList={siteMediaList}/>}
-      {hasQuestions && <QuestionEditor section={section} updateSection={updateSection}/>}
-      {hasLogos && <LogoEditor portalEnvContext={portalEnvContext} section={section}
-        updateSection={updateSection} siteMediaList={siteMediaList}/>}
-      {hasButtons && <ButtonEditor section={section} updateSection={updateSection}/>}
-      {hasSubGrids && <PhotoBioEditor portalEnvContext={portalEnvContext} mediaList={siteMediaList}
-        section={section} updateSection={updateSection}/>}
+      {hasTitle &&
+          <TitleEditor section={section} updateSection={updateSection}/>}
+      {hasBlurb &&
+          <BlurbEditor section={section} updateSection={updateSection}/>}
+      {hasImage &&
+          <ImageEditor portalEnvContext={portalEnvContext} section={section}
+            updateSection={updateSection} siteMediaList={siteMediaList}/>}
+      {/* All sections have a style editor */}
+      <StyleEditor section={section} updateSection={updateSection}/>
+      {hasSteps &&
+          <ParticipationStepsEditor portalEnvContext={portalEnvContext} section={section}
+            updateSection={updateSection} siteMediaList={siteMediaList}/>}
+      {hasQuestions &&
+          <FrequentlyAskedQuestionEditor section={section} updateSection={updateSection}/>}
+      {hasLogos &&
+          <LogoEditor portalEnvContext={portalEnvContext} section={section}
+            updateSection={updateSection} siteMediaList={siteMediaList}/>}
+      {hasButtons &&
+          <ButtonEditor section={section} updateSection={updateSection}/>}
+      {hasSubGrids &&
+          <PhotoBioEditor portalEnvContext={portalEnvContext} mediaList={siteMediaList}
+            section={section} updateSection={updateSection}/>}
     </div>
   )
 }
@@ -345,7 +224,7 @@ const SectionEditorComponents: Record<SectionType, SectionEditorComponentType | 
   FAQ: DynamicSectionEditor,
   HERO_CENTERED: DynamicSectionEditor,
   HERO_WITH_IMAGE: DynamicSectionEditor,
-  SOCIAL_MEDIA: SocialMediaSectionEditor,
+  SOCIAL_MEDIA: SocialMediaEditor,
   STEP_OVERVIEW: DynamicSectionEditor,
   PHOTO_BLURB_GRID: DynamicSectionEditor,
   PARTICIPATION_DETAIL: DynamicSectionEditor,
@@ -353,511 +232,3 @@ const SectionEditorComponents: Record<SectionType, SectionEditorComponentType | 
   LINK_SECTIONS_FOOTER: DynamicSectionEditor,
   BANNER_IMAGE: DynamicSectionEditor
 }
-
-
-/* UPDATED COMPONENTS DOWN HERE */
-
-const TitleEditor = ({ section, updateSection }: {
-  section: HtmlSection, updateSection: (section: HtmlSection) => void
-}) => {
-  const config = JSON.parse(section.sectionConfig || '{}') as SectionConfig
-  return (
-    <TextInput className="mb-2" label="Title" value={config.title as string}
-      placeholder={'Enter a title for this section'}
-      onChange={value => {
-        const parsed = JSON.parse(section.sectionConfig || '{}')
-        updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, title: value }) })
-      }}/>
-  )
-}
-
-const BlurbEditor = ({ section, updateSection }: {
-  section: HtmlSection, updateSection: (section: HtmlSection) => void
-}) => {
-  const config = JSON.parse(section.sectionConfig || '{}') as SectionConfig
-  return (
-    <Textarea rows={3} className="mb-2" label="Blurb" value={config.blurb as string}
-      placeholder={'Enter a blurb for this section'}
-      onChange={value => {
-        const parsed = JSON.parse(section.sectionConfig || '{}')
-        updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, blurb: value }) })
-      }}/>
-  )
-}
-
-const QuestionEditor = ({ section, updateSection }: {
-  section: HtmlSection, updateSection: (section: HtmlSection) => void
-}) => {
-  const config = JSON.parse(section.sectionConfig || '{}') as SectionConfig
-  const questions = config.questions as FaqQuestion[] || []
-  const faqContentId = useId()
-  const faqTargetSelector = `#${faqContentId}`
-  return (<div>
-    <div className="pb-1">
-      <button
-        aria-controls={faqTargetSelector}
-        aria-expanded="true"
-        className={classNames('btn w-100 py-2 px-0 d-flex text-decoration-none')}
-        data-bs-target={faqTargetSelector}
-        data-bs-toggle="collapse"
-      >
-        <span className={'form-label fw-semibold mb-0'}>Questions ({questions.length})</span>
-        <span className="text-center px-2">
-          <FontAwesomeIcon icon={faChevronDown} className="hidden-when-collapsed"/>
-          <FontAwesomeIcon icon={faChevronUp} className="hidden-when-expanded"/>
-        </span>
-      </button>
-    </div>
-    <div className="collapse hide rounded-3 mb-2" id={faqContentId}
-      style={{ backgroundColor: '#eee', padding: '0.75rem' }}>
-      {questions.map((question, i) => {
-        return <div key={i} style={{ backgroundColor: '#ddd', padding: '0.75rem' }} className="rounded-3 mb-2">
-          <div className="d-flex justify-content-between align-items-center">
-            <span className="h5">Question {i + 1}</span>
-            <div role="button" className="d-flex justify-content-end">
-              <FontAwesomeIcon icon={faTimes} className={'text-danger'} onClick={() => {
-                const parsed = JSON.parse(section.sectionConfig!)
-                const newQuestions = [...config.questions as FaqQuestion[]]
-                newQuestions.splice(i, 1)
-                updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, questions: newQuestions }) })
-              }}/>
-            </div>
-          </div>
-          <TextInput label="Question" value={question.question} onChange={value => {
-            const parsed = JSON.parse(section.sectionConfig!)
-            const newQuestions = [...config.questions as FaqQuestion[]]
-            newQuestions[i].question = value
-            updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, questions: newQuestions }) })
-          }}/>
-          <Textarea rows={3} label="Answer" value={question.answer} onChange={value => {
-            const parsed = JSON.parse(section.sectionConfig!)
-            const newQuestions = [...config.questions as FaqQuestion[]]
-            newQuestions[i].answer = value
-            updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, questions: newQuestions }) })
-          }}/>
-        </div>
-      })}
-      <Button onClick={() => {
-        const parsed = JSON.parse(section.sectionConfig!)
-        const newQuestions = [...config.questions as FaqQuestion[]]
-        newQuestions.push({ question: '', answer: '' })
-        updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, questions: newQuestions }) })
-      }}><FontAwesomeIcon icon={faPlus}/> Add Question</Button>
-    </div>
-  </div>
-  )
-}
-
-const StepEditor = ({ portalEnvContext, section, updateSection, siteMediaList }: {
-  portalEnvContext: PortalEnvContext,
-  section: HtmlSection, updateSection: (section: HtmlSection) => void, siteMediaList: SiteMediaMetadata[]
-}) => {
-  const config = JSON.parse(section.sectionConfig || '{}') as SectionConfig
-  const steps = config.steps as StepConfig[] || []
-  const stepContentId = useId()
-  const stepTargetSelector = `#${stepContentId}`
-  return (
-    <div>
-      <div className="pb-1">
-        <button
-          aria-controls={stepTargetSelector}
-          aria-expanded="true"
-          className={classNames('btn w-100 py-2 px-0 d-flex text-decoration-none')}
-          data-bs-target={stepTargetSelector}
-          data-bs-toggle="collapse"
-        >
-          <span className={'form-label fw-semibold mb-0'}>Steps ({steps.length})</span>
-          <span className="text-center px-2">
-            <FontAwesomeIcon icon={faChevronDown} className="hidden-when-collapsed"/>
-            <FontAwesomeIcon icon={faChevronUp} className="hidden-when-expanded"/>
-          </span>
-        </button>
-      </div>
-      <div className="collapse hide rounded-3 mb-2" id={stepContentId}
-        style={{ backgroundColor: '#eee', padding: '0.75rem' }}>
-        <div className='mb-2'><Checkbox label={'Show Step Numbers'}
-          checked={config.showStepNumbers as boolean == undefined ? true : config.showStepNumbers as boolean}
-          onChange={value => {
-            const parsed = JSON.parse(section.sectionConfig || '{}')
-            updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, showStepNumbers: value }) })
-          }}/></div>
-        <div>
-          {steps.map((step, i) => {
-            return <div key={i} style={{ backgroundColor: '#ddd', padding: '0.75rem' }} className="rounded-3 mb-2">
-              <div className="d-flex justify-content-between align-items-center">
-                <span className="h5">Step {i + 1}</span>
-                <div role="button" className="d-flex justify-content-end">
-                  <FontAwesomeIcon icon={faTimes} className={'text-danger'} onClick={() => {
-                    const parsed = JSON.parse(section.sectionConfig!)
-                    const newSteps = [...config.steps as StepConfig[]]
-                    newSteps.splice(i, 1)
-                    updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, steps: newSteps }) })
-                  }}/></div>
-              </div>
-              <div>
-                <label className='form-label fw-semibold m-0'>Image</label>
-                <ImageSelector portalEnvContext={portalEnvContext}
-                  imageList={siteMediaList} image={step.image as ImageConfig} onChange={image => {
-                    const parsed = JSON.parse(section.sectionConfig!)
-                    const newSteps = [...config.steps as StepConfig[]]
-                    newSteps[i].image = image
-                    updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, steps: newSteps }) })
-                  }}/>
-                <TextInput label="Duration" value={step.duration} onChange={value => {
-                  const parsed = JSON.parse(section.sectionConfig!)
-                  const newSteps = [...config.steps as StepConfig[]]
-                  newSteps[i].duration = value
-                  updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, steps: newSteps }) })
-                }}/>
-                <Textarea rows={2} label="Blurb" value={step.blurb} onChange={value => {
-                  const parsed = JSON.parse(section.sectionConfig!)
-                  const newSteps = [...config.steps as StepConfig[]]
-                  newSteps[i].blurb = value
-                  updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, steps: newSteps }) })
-                }}/>
-              </div>
-            </div>
-          })}
-        </div>
-        <Button onClick={() => {
-          const parsed = JSON.parse(section.sectionConfig!)
-          const newSteps = [...config.steps as StepConfig[]]
-          newSteps.push({ image: { cleanFileName: '', version: 1 }, duration: '', blurb: '' })
-          updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, steps: newSteps }) })
-        }}><FontAwesomeIcon icon={faPlus}/> Add Step</Button>
-      </div>
-    </div>
-  )
-}
-
-const ImageEditor = ({ portalEnvContext, section, updateSection, siteMediaList }: {
-  portalEnvContext: PortalEnvContext,
-  section: HtmlSection, updateSection: (section: HtmlSection) => void, siteMediaList: SiteMediaMetadata[]
-}) => {
-  const config = JSON.parse(section.sectionConfig || '{}') as SectionConfig
-  return (
-    <div>
-      <label className='form-label fw-semibold'>Image</label>
-      <ImageSelector portalEnvContext={portalEnvContext}
-        imageList={siteMediaList} image={config.image as ImageConfig} onChange={image => {
-          updateSection({ ...section, sectionConfig: JSON.stringify({ ...config, image }) })
-        }}/>
-    </div>
-  )
-}
-
-const LogoEditor = ({ portalEnvContext, section, updateSection, siteMediaList }: {
-  portalEnvContext: PortalEnvContext,
-  section: HtmlSection, updateSection: (section: HtmlSection) => void, siteMediaList: SiteMediaMetadata[]
-}) => {
-  const config = JSON.parse(section.sectionConfig || '{}') as SectionConfig
-  const logos = config.logos as ImageConfig[] || []
-
-  const logosContentId = useId()
-  const logosTargetSelector = `#${logosContentId}`
-  return (
-    <div>
-      <div className="pb-1">
-        <button
-          aria-controls={logosTargetSelector}
-          aria-expanded="true"
-          className={classNames('btn w-100 py-2 px-0 d-flex text-decoration-none')}
-          data-bs-target={logosTargetSelector}
-          data-bs-toggle="collapse"
-        >
-          <span className={'form-label fw-semibold mb-0'}>Logos ({logos.length})</span>
-          <span className="text-center px-2">
-            <FontAwesomeIcon icon={faChevronDown} className="hidden-when-collapsed"/>
-            <FontAwesomeIcon icon={faChevronUp} className="hidden-when-expanded"/>
-          </span>
-        </button>
-      </div>
-      <div className="collapse hide rounded-3 mb-2" id={logosContentId}
-        style={{ backgroundColor: '#eee', padding: '0.75rem' }}>
-        <div>
-          {logos.map((logo, i) => {
-            return <div key={i} style={{ backgroundColor: '#ddd', padding: '0.75rem' }} className="rounded-3 mb-2">
-              <div className="d-flex justify-content-between align-items-center">
-                <span className="h5">Logo {i + 1}</span>
-                <div role="button" className="d-flex justify-content-end">
-                  <FontAwesomeIcon icon={faTimes} className={'text-danger'} onClick={() => {
-                    const parsed = JSON.parse(section.sectionConfig!)
-                    const newLogos = [...config.logos as MediaConfig[]]
-                    newLogos.splice(i, 1)
-                    updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, logos: newLogos }) })
-                  }}/></div>
-              </div>
-              <div>
-                <label className='form-label fw-semibold m-0'>Image</label>
-                <ImageSelector portalEnvContext={portalEnvContext}
-                  imageList={siteMediaList} image={logo as ImageConfig} onChange={image => {
-                    const parsed = JSON.parse(section.sectionConfig || '{}')
-                    const newLogos = [...config.logos as MediaConfig[]]
-                    newLogos[i] = {
-                      cleanFileName: image.cleanFileName,
-                      version: image.version
-                    }
-                    updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, logos: newLogos }) })
-                  }}/>
-                <TextInput label="Alt Text" value={logo.alt} onChange={value => {
-                  const parsed = JSON.parse(section.sectionConfig || '{}')
-                  const newLogos = [...config.logos as MediaConfig[]]
-                  newLogos[i].alt = value
-                  updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, logos: newLogos }) })
-                }}/>
-
-                <TextInput label="Link" value={logo.link} onChange={value => {
-                  const parsed = JSON.parse(section.sectionConfig || '{}')
-                  const newLogos = [...config.logos as ImageConfig[]]
-                  newLogos[i].link = value
-                  updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, logos: newLogos }) })
-                }}/>
-
-              </div>
-            </div>
-          })}
-        </div>
-        <Button onClick={() => {
-          const parsed = JSON.parse(section.sectionConfig!)
-          const newLogos = [...logos]
-          newLogos.push({ cleanFileName: '', version: 1 })
-          updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, logos: newLogos }) })
-        }}><FontAwesomeIcon icon={faPlus}/> Add Logo</Button>
-      </div>
-    </div>
-  )
-}
-
-const PhotoBioEditor = ({ portalEnvContext, mediaList, section, updateSection }: {
-  portalEnvContext: PortalEnvContext, mediaList: SiteMediaMetadata[],
-  section: HtmlSection, updateSection: (section: HtmlSection) => void
-}) => {
-  const config = JSON.parse(section.sectionConfig || '{}') as SectionConfig
-  const subGrids = config.subGrids as SubGrid[] || []
-
-  const photosContentId = useId()
-  const photosTargetSelector = `#${photosContentId}`
-  return (
-    <div>
-      <div className="pb-1">
-        <button
-          aria-controls={photosTargetSelector}
-          aria-expanded="true"
-          className={classNames('btn w-100 py-2 px-0 d-flex text-decoration-none')}
-          data-bs-target={photosTargetSelector}
-          data-bs-toggle="collapse"
-        >
-          <span className={'form-label fw-semibold mb-0'}>Photo Bio Sections ({subGrids.length})</span>
-          <span className="text-center px-2">
-            <FontAwesomeIcon icon={faChevronDown} className="hidden-when-collapsed"/>
-            <FontAwesomeIcon icon={faChevronUp} className="hidden-when-expanded"/>
-          </span>
-        </button>
-      </div>
-      <div className="collapse hide rounded-3 mb-2" id={photosContentId}
-        style={{ backgroundColor: '#eee', padding: '0.75rem' }}>
-        <div>
-          {subGrids.map((subGrid, i) => {
-            return <div key={i} style={{ backgroundColor: '#ddd', padding: '0.75rem' }} className="rounded-3 mb-2">
-              <div className="d-flex justify-content-between align-items-center">
-                {/*<span className="h5">Photo Bio {i + 1}</span>*/}
-                <PhotoBioSectionEditor portalEnvContext={portalEnvContext}
-                  index={i}
-                  mediaList={mediaList} section={section} updateSection={updateSection}/>
-              </div>
-            </div>
-          })}
-        </div>
-      </div>
-    </div>
-  )
-}
-
-const PhotoBioSectionEditor = ({ portalEnvContext, index, mediaList, section, updateSection }: {
-    portalEnvContext: PortalEnvContext,
-  index: number, mediaList: SiteMediaMetadata[], section: HtmlSection, updateSection: (section: HtmlSection) => void
-}) => {
-  const config = JSON.parse(section.sectionConfig || '{}') as SectionConfig
-  const subGrids = config.subGrids as SubGrid[] || []
-  const subGrid = subGrids[index]
-
-  const subGridsContentId = useId()
-  const subGridsTargetSelector = `#${subGridsContentId}`
-  return (
-    <div className='w-100'>
-      <div className="pb-1">
-        <button
-          aria-controls={subGridsTargetSelector}
-          aria-expanded="true"
-          className={classNames('btn w-100 py-2 px-0 d-flex text-decoration-none')}
-          data-bs-target={subGridsTargetSelector}
-          data-bs-toggle="collapse"
-        >
-          <span className={'form-label fw-semibold mb-0'}>
-            {subGrid.title ? subGrid.title : 'Bio sub-section'} ({subGrid.photoBios.length})
-          </span> <span className="text-center px-2">
-            <FontAwesomeIcon icon={faChevronDown} className="hidden-when-collapsed"/>
-            <FontAwesomeIcon icon={faChevronUp} className="hidden-when-expanded"/>
-          </span>
-        </button>
-      </div>
-      <div className="collapse hide rounded-3 mb-2" id={subGridsContentId}
-        style={{ backgroundColor: '#eee', padding: '0.75rem' }}>
-        <div>
-          {subGrid.photoBios.map((photoBio, i) => {
-            return <div key={i} style={{ backgroundColor: '#ddd', padding: '0.75rem' }} className="rounded-3 mb-2">
-              <div>
-                <label className='form-label fw-semibold m-0'>Image</label>
-                <ImageSelector portalEnvContext={portalEnvContext}
-                  imageList={mediaList} image={photoBio.image as ImageConfig} onChange={image => {
-                    const parsed = JSON.parse(section.sectionConfig || '{}')
-                    const newSubGrids = [...config.subGrids as SectionConfig[]]
-                    newSubGrids[i].image = image
-                    updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, subGrids: newSubGrids }) })
-                  }}/>
-                <TextInput label="Name" value={photoBio.name} onChange={value => {
-                  const parsed = JSON.parse(section.sectionConfig || '{}')
-                  const newSubGrids = [...config.subGrids as SectionConfig[]]
-                  newSubGrids[i].name = value
-                  updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, subGrids: newSubGrids }) })
-                }}/>
-                <TextInput label="Title" value={photoBio.title} onChange={value => {
-                  const parsed = JSON.parse(section.sectionConfig || '{}')
-                  const newSubGrids = [...config.subGrids as SectionConfig[]]
-                  newSubGrids[i].title = value
-                  updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, subGrids: newSubGrids }) })
-                }}/>
-                <Textarea rows={2} label="Bio" value={photoBio.blurb} onChange={value => {
-                  const parsed = JSON.parse(section.sectionConfig || '{}')
-                  const newSubGrids = [...config.subGrids as SectionConfig[]]
-                  newSubGrids[i].bio = value
-                  updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, subGrids: newSubGrids }) })
-                }}/>
-                <Textarea rows={2} label="Detail" value={photoBio.detail} onChange={value => {
-                  const parsed = JSON.parse(section.sectionConfig || '{}')
-                  const newSubGrids = [...config.subGrids as SectionConfig[]]
-                  newSubGrids[i].detail = value
-                  updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, subGrids: newSubGrids }) })
-                }}/>
-              </div>
-            </div>
-          })}
-        </div>
-      </div>
-    </div>
-  )
-}
-
-const ButtonEditor = ({ section, updateSection }: {
-    section: HtmlSection, updateSection: (section: HtmlSection) => void
-    }) => {
-  const config = JSON.parse(section.sectionConfig || '{}') as SectionConfig
-  const buttons = config.buttons as ButtonConfig[] || []
-  const buttonsContentId = useId()
-  const buttonsTargetSelector = `#${buttonsContentId}`
-  return (
-    <div>
-      <div className="pb-1">
-        <button
-          aria-controls={buttonsTargetSelector}
-          aria-expanded="true"
-          className={classNames('btn w-100 py-2 px-0 d-flex text-decoration-none')}
-          data-bs-target={buttonsTargetSelector}
-          data-bs-toggle="collapse"
-        >
-          <span className={'form-label fw-semibold mb-0'}>Buttons ({buttons.length})</span>
-          <span className="text-center px-2">
-            <FontAwesomeIcon icon={faChevronDown} className="hidden-when-collapsed"/>
-            <FontAwesomeIcon icon={faChevronUp} className="hidden-when-expanded"/>
-          </span>
-        </button>
-      </div>
-      <div className="collapse hide rounded-3 mb-2" id={buttonsContentId}
-        style={{ backgroundColor: '#eee', padding: '0.75rem' }}>
-        <div>
-          {buttons.map((button, i) => {
-            return <div key={i} style={{ backgroundColor: '#ddd', padding: '0.75rem' }} className="rounded-3 mb-2">
-              <div className="d-flex justify-content-between align-items-center">
-                <span className="h5">Button {i + 1}</span>
-                <div role="button" className="d-flex justify-content-end">
-                  <FontAwesomeIcon icon={faTimes} className={'text-danger'} onClick={() => {
-                    const parsed = JSON.parse(section.sectionConfig!)
-                    const newButtons = [...config.buttons as ButtonConfig[]]
-                    newButtons.splice(i, 1)
-                    updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, buttons: newButtons }) })
-                  }}/></div>
-              </div>
-              <div>
-                <label className='form-label fw-semibold m-0'>Button Type</label>
-                <Select className='w-100' options={BUTTON_TYPES} value={BUTTON_TYPES[
-                  BUTTON_TYPES.findIndex(opt => opt.value === button.type)
-                ]} aria-label={'Select section type'}
-                onChange={opt => {
-                  if (opt != undefined) {
-                    const parsed = JSON.parse(section.sectionConfig || '{}')
-                    const newButtons = [...config.buttons as ButtonConfig[]]
-                    newButtons[i].type = ButtonTypeTemplates[opt.value].type
-                    updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, buttons: newButtons }) })
-                  }
-                }}/>
-                <TextInput label="Button Text" value={button.text} onChange={value => {
-                  const parsed = JSON.parse(section.sectionConfig || '{}')
-                  const newButtons = [...config.buttons as ButtonConfig[]]
-                  newButtons[i].text = value
-                  updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, buttons: newButtons }) })
-                }}/>
-                {/* @ts-ignore */}
-                {button.type === 'externalLink' || button.type === 'internalLink' &&
-                  <TextInput label="Button Link" value={button.href} onChange={value => {
-                    const parsed = JSON.parse(section.sectionConfig || '{}')
-                    const newButtons = [...config.buttons as ButtonConfig[]]
-                    // @ts-ignore
-                    newButtons[i].href = value
-                    updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, buttons: newButtons }) })
-                  }}/>}
-                {button.type === 'join' &&
-                  <TextInput label="Study Shortcode" value={button.studyShortcode} onChange={value => {
-                    const parsed = JSON.parse(section.sectionConfig || '{}')
-                    const newButtons = [...config.buttons as ButtonConfig[]]
-                    // @ts-ignore
-                    newButtons[i].studyShortcode = value
-                    updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, buttons: newButtons }) })
-                  }}/>}
-              </div>
-            </div>
-          })}
-        </div>
-        <Button onClick={() => {
-          const parsed = JSON.parse(section.sectionConfig!)
-          const newButtons = [...buttons]
-          newButtons.push({ type: undefined, text: '', href: '' })
-          updateSection({ ...section, sectionConfig: JSON.stringify({ ...parsed, buttons: newButtons }) })
-        }}><FontAwesomeIcon icon={faPlus}/> Add Button
-        </Button>
-      </div>
-    </div>
-  )
-}
-
-const ButtonTypeTemplates: Record<string, ButtonConfig> = {
-  'internalLink': {
-    type: 'internalLink',
-    text: '',
-    href: ''
-  },
-  'join': {
-    type: 'join',
-    text: 'Join Study',
-    studyShortcode: ''
-  },
-  'mailingList': {
-    type: 'mailingList',
-    text: 'Join Mailing List'
-  },
-  'externalLink': {
-    type: undefined,
-    text: '',
-    href: ''
-  }
-}
-
