+++ ui-core/src/index.ts
@@ -39,6 +39,8 @@ export * from './participant/landing/sections/FrequentlyAskedQuestionsTemplate'
 export * from './participant/landing/sections/SocialMediaTemplate'
 export * from './participant/landing/sections/HeroCenteredTemplate'
 export * from './participant/landing/sections/PhotoBlurbGrid'
+export * from './participant/landing/ConfiguredButton'
+export * from './participant/landing/ConfiguredMedia'
 export * from './participant/landing/Markdown'
 export * from './participant/landing/MailingListModal'
 export * from './participant/landing/MailingListForm'
